Partial updates
Support a variety of partial updates:
- Update one or more properties on the top-level document
- Update one or more properties in embedded resources
- Update entire embedded resource
- Update entire embedded collection

At the moment, specific properties in embedded collections can't be
updated - largely because we can't assume consistent ordering of
elements in the collection source, so we can't guarantee that the
collection source elements map in order to the saved collection
elements.

All specs are passing but the code could most likely use some cleanup.